refactor: move computeShares and saveDelegatorSnapshot to helpers
These follow the existing patterns in helpers.ts (percOf/precisePercOf
for math, createOrLoad* for entity creation).
